fix(bridge): Wisher is *uint16 — Magic_lamp wisher 256 no longer drops makes
Magic_lamp's `wisher` is UNASSIGNED_NOID (256) when no wish is in flight
(Magic_lamp.java), which overflowed the *uint8 schema field — json.Unmarshal
then failed ("cannot unmarshal number 256 into ... uint8") and the bridge
dropped the whole `make` message (the library genie never appeared). Widen
Wisher to *uint16, matching the existing Appearing/Speaker convention:

  - schema.go: Wisher *uint8 -> *uint16
  - client_session.go: rewriteNestedNoid16 for the uint16 field (skips the 256
    sentinel; remaps real avatar noids on session restore)
  - elko_state_encoders.go: encode via noidU8 (256 -> GHOST_NOID) like Appearing
  - test: wisher var uint16

go build / vet / full bridge suite green.

+++ bridge_v2/bridge/restore_noid_rewrite_test.go
@@ -177,7 +177,7 @@ func TestRestore_RewritesWisherUsingPriorMapping(t *testing.T) {
 	}
 
 	lampMsg := makeMsg("Magic_lamp", "item-lamp-1", 60)
-	wisher := uint8(31) // elko's noid for randy
+	wisher := uint16(31) // elko's noid for randy
 	lampMsg.Obj.Mods[0].Wisher = &wisher
 	if err := sess.unpackHabitatObject(lampMsg, "context-Downtown_5c"); err != nil {
 		t.Fatalf("lamp unpack: %v", err)
